refactor: move runner workspace sync types to runner-contract (stage 4)
Continues decoupling the optional runner feature from core. Moves two more
behavior-free contract types out of runner/workspace/types.rs into
homeboy-runner-contract:
- RunnerWorkspaceSyncMode (enum + its pure label() method)
- RunnerWorkspaceSyncOptions

Both are plain data (fields are primitives + each other), re-exported from
runner::workspace::types so internal/CLI call sites resolve unchanged. External
consumers (upgrade/runners, daemon/patch_capture, agent_task_scheduler/harvest)
repointed to homeboy_runner_contract::, severing those reverse edges.

Note: RunnerExecOptions was the original stage-4 target but it turned out to be
a ~30-method behavior builder that depends on core's SourceSnapshot — it's a
behavioral (Category B / hook-invert) edge, not a movable type. Corrected the
plan accordingly; picked these clean plain-data types instead.

Part of the staged runner extraction plan (wiki: homeboy-runner-crate-
extraction-plan).

Verified: cargo check clean for -p homeboy-runner-contract, -p homeboy-core
--lib, -p homeboy-cli --lib.
